Refactor how linux/C port initialize the serial fd
This is to bring the ported library in line with the arduino C++ version
where the user is expected to initialize the serial connection and then
pass it to the xmodem send/receive functions.

Since configuring the tty connection can be quite complicated it
initially made sense for the library to provide this function but
eventually I decided that keeping parity in the usage patterns was more
important even though it meant adding complicated initialization code to
all the examples.